// lib/types.ts
export type KaffahData = {
      id: string;
      tahun: string;
      name: string;
      januari: string;
      februari: string;
      maret: string;
      april: string;
      mei: string;
      juni: string;
      juli: string;
      agustus: string;
      september: string;
      oktober: string;
      november: string;
      desember: string;
}